+++ spec/rubocop/cop/rails/find_each_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Rails::FindEach do
-  subject(:cop) { described_class.new }
+RSpec.describe RuboCop::Cop::Rails::FindEach, :config do
+  subject(:cop) { described_class.new(config) }
 
   shared_examples 'register_offense' do |scope|
     it "registers an offense when using #{scope}.each" do
@@ -67,7 +67,22 @@ RSpec.describe RuboCop::Cop::Rails::FindEach do
     RUBY
   end
 
-  it 'does not register an offense when using order(...) earlier' do
-    expect_no_offenses('User.order(:name).all.each { |u| u.something }')
+  context 'ignored methods' do
+    let(:cop_config) { { 'IgnoredMethods' => %w[order lock] } }
+
+    it 'does not register an offense when using order(...) earlier' do
+      expect_no_offenses('User.order(:name).each { |u| u.something }')
+    end
+
+    it 'does not register an offense when using lock earlier' do
+      expect_no_offenses('User.lock.each { |u| u.something }')
+    end
+
+    it 'registers offense for methods not in `IgnoredMethods`' do
+      expect_offense(<<~RUBY)
+        User.joins(:posts).each { |u| u.something }
+                           ^^^^ Use `find_each` instead of `each`.
+      RUBY
+    end
   end
 end